chore: move validator_rules to helper folder - enable modern syntax for importing helpers in package.json
