+++ app/controllers/auth_controller.ts
@@ -4,7 +4,7 @@ import hash from '@adonisjs/core/services/hash'
 import { schema } from '@adonisjs/validator'
 import { generateOtp } from '../../helpers/utils.js'
 import router from '@adonisjs/core/services/router'
-import { rules } from '#services/validator_rules'
+import { rules } from '#helpers/validator_rules'
 
 export default class AuthController {
   /**
